requirements.txt
disnake
python-dotenv
pillow
pymongo
matplotlib
numpy